Extract webhook repository logic out of the core github repository
