+++ pkg/registry/apis/provisioning/repository/github.go
@@ -8,11 +8,8 @@ import (
 	"net/http"
 	"net/url"
 	"regexp"
-	"slices"
 	"strings"
 
-	"github.com/google/go-github/v70/github"
-	"github.com/google/uuid"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/validation/field"
@@ -28,10 +25,9 @@ var subscribedEvents = []string{"push", "pull_request"}
 
 // Make sure all public functions of this struct call the (*githubRepository).logger function, to ensure the GH repo details are included.
 type githubRepository struct {
-	config     *provisioning.Repository
-	gh         pgh.Client // assumes github.com base URL
-	secrets    secrets.Service
-	webhookURL string
+	config  *provisioning.Repository
+	gh      pgh.Client // assumes github.com base URL
+	secrets secrets.Service
 
 	owner string
 	repo  string
@@ -39,24 +35,27 @@ type githubRepository struct {
 	cloneFn CloneFn
 }
 
-var (
-	_ Repository         = (*githubRepository)(nil)
-	_ Hooks              = (*githubRepository)(nil)
-	_ Versioned          = (*githubRepository)(nil)
-	_ Writer             = (*githubRepository)(nil)
-	_ Reader             = (*githubRepository)(nil)
-	_ RepositoryWithURLs = (*githubRepository)(nil)
-	_ ClonableRepository = (*githubRepository)(nil)
-)
+// GithubRepository is an interface that combines all repository capabilities
+// needed for GitHub repositories.
+type GithubRepository interface {
+	Repository
+	Versioned
+	Writer
+	Reader
+	RepositoryWithURLs
+	ClonableRepository
+	Owner() string
+	Repo() string
+	Client() pgh.Client
+}
 
 func NewGitHub(
 	ctx context.Context,
 	config *provisioning.Repository,
 	factory *pgh.Factory,
 	secrets secrets.Service,
-	webhookURL string,
 	cloneFn CloneFn,
-) (*githubRepository, error) {
+) (GithubRepository, error) {
 	owner, repo, err := parseOwnerRepo(config.Spec.GitHub.URL)
 	if err != nil {
 		return nil, fmt.Errorf("parse owner and repo: %w", err)
@@ -72,18 +71,29 @@ func NewGitHub(
 	}
 
 	return &githubRepository{
-		config:     config,
-		gh:         factory.New(ctx, token), // TODO, baseURL from config
-		secrets:    secrets,
-		webhookURL: webhookURL,
-		owner:      owner,
-		repo:       repo,
-		cloneFn:    cloneFn,
+		config:  config,
+		gh:      factory.New(ctx, token), // TODO, baseURL from config
+		secrets: secrets,
+		owner:   owner,
+		repo:    repo,
+		cloneFn: cloneFn,
 	}, nil
 }
 
 func (r *githubRepository) Config() *provisioning.Repository {
 	return r.config
+}
+
+func (r *githubRepository) Owner() string {
+	return r.owner
+}
+
+func (r *githubRepository) Repo() string {
+	return r.repo
+}
+
+func (r *githubRepository) Client() pgh.Client {
+	return r.gh
 }
 
 // Validate implements provisioning.Repository.
@@ -512,130 +522,6 @@ func (r *githubRepository) ensureBranchExists(ctx context.Context, branchName st
 	return nil
 }
 
-// Webhook implements Repository.
-func (r *githubRepository) Webhook(ctx context.Context, req *http.Request) (*provisioning.WebhookResponse, error) {
-	if r.config.Status.Webhook == nil {
-		return nil, fmt.Errorf("unexpected webhook request")
-	}
-
-	secret, err := r.secrets.Decrypt(ctx, r.config.Status.Webhook.EncryptedSecret)
-	if err != nil {
-		return nil, fmt.Errorf("failed to decrypt secret: %w", err)
-	}
-
-	payload, err := github.ValidatePayload(req, secret)
-	if err != nil {
-		return nil, apierrors.NewUnauthorized("invalid signature")
-	}
-
-	return r.parseWebhook(github.WebHookType(req), payload)
-}
-
-// This method does not include context because it does delegate any more requests
-func (r *githubRepository) parseWebhook(messageType string, payload []byte) (*provisioning.WebhookResponse, error) {
-	event, err := github.ParseWebHook(messageType, payload)
-	if err != nil {
-		return nil, apierrors.NewBadRequest("invalid payload")
-	}
-
-	switch event := event.(type) {
-	case *github.PushEvent:
-		return r.parsePushEvent(event)
-	case *github.PullRequestEvent:
-		return r.parsePullRequestEvent(event)
-	case *github.PingEvent:
-		return &provisioning.WebhookResponse{
-			Code:    http.StatusOK,
-			Message: "ping received",
-		}, nil
-	default:
-		return &provisioning.WebhookResponse{
-			Code:    http.StatusNotImplemented,
-			Message: fmt.Sprintf("unsupported messageType: %s", messageType),
-		}, nil
-	}
-}
-
-func (r *githubRepository) parsePushEvent(event *github.PushEvent) (*provisioning.WebhookResponse, error) {
-	if event.GetRepo() == nil {
-		return nil, fmt.Errorf("missing repository in push event")
-	}
-	if event.GetRepo().GetFullName() != fmt.Sprintf("%s/%s", r.owner, r.repo) {
-		return nil, fmt.Errorf("repository mismatch")
-	}
-
-	// No need to sync if not enabled
-	if !r.config.Spec.Sync.Enabled {
-		return &provisioning.WebhookResponse{Code: http.StatusOK}, nil
-	}
-
-	// Skip silently if the event is not for the main/master branch
-	// as we cannot configure the webhook to only publish events for the main branch
-	if event.GetRef() != fmt.Sprintf("refs/heads/%s", r.config.Spec.GitHub.Branch) {
-		return &provisioning.WebhookResponse{Code: http.StatusOK}, nil
-	}
-
-	return &provisioning.WebhookResponse{
-		Code: http.StatusAccepted,
-		Job: &provisioning.JobSpec{
-			Repository: r.Config().GetName(),
-			Action:     provisioning.JobActionPull,
-			Pull: &provisioning.SyncJobOptions{
-				Incremental: true,
-			},
-		},
-	}, nil
-}
-
-func (r *githubRepository) parsePullRequestEvent(event *github.PullRequestEvent) (*provisioning.WebhookResponse, error) {
-	if event.GetRepo() == nil {
-		return nil, fmt.Errorf("missing repository in pull request event")
-	}
-	cfg := r.config.Spec.GitHub
-	if cfg == nil {
-		return nil, fmt.Errorf("missing GitHub config")
-	}
-
-	if event.GetRepo().GetFullName() != fmt.Sprintf("%s/%s", r.owner, r.repo) {
-		return nil, fmt.Errorf("repository mismatch")
-	}
-	pr := event.GetPullRequest()
-	if pr == nil {
-		return nil, fmt.Errorf("expected PR in event")
-	}
-
-	if pr.GetBase().GetRef() != r.config.Spec.GitHub.Branch {
-		return &provisioning.WebhookResponse{
-			Code:    http.StatusOK,
-			Message: fmt.Sprintf("ignoring pull request event as %s is not  the configured branch", pr.GetBase().GetRef()),
-		}, nil
-	}
-
-	action := event.GetAction()
-	if action != "opened" && action != "reopened" && action != "synchronize" {
-		return &provisioning.WebhookResponse{
-			Code:    http.StatusOK, // Nothing needed
-			Message: fmt.Sprintf("ignore pull request event: %s", action),
-		}, nil
-	}
-
-	// Queue an async job that will parse files
-	return &provisioning.WebhookResponse{
-		Code:    http.StatusAccepted, // Nothing needed
-		Message: fmt.Sprintf("pull request: %s", action),
-		Job: &provisioning.JobSpec{
-			Repository: r.Config().GetName(),
-			Action:     provisioning.JobActionPullRequest,
-			PullRequest: &provisioning.PullRequestJobOptions{
-				URL:  pr.GetHTMLURL(),
-				PR:   pr.GetNumber(),
-				Ref:  pr.GetHead().GetRef(),
-				Hash: pr.GetHead().GetSHA(),
-			},
-		},
-	}, nil
-}
-
 func (r *githubRepository) LatestRef(ctx context.Context) (string, error) {
 	ctx, _ = r.logger(ctx, "")
 	branch, err := r.gh.GetBranch(ctx, r.owner, r.repo, r.Config().Spec.GitHub.Branch)
@@ -746,12 +632,6 @@ func (r *githubRepository) CompareFiles(ctx context.Context, base, ref string) (
 	return changes, nil
 }
 
-// CommentPullRequest adds a comment to a pull request.
-func (r *githubRepository) CommentPullRequest(ctx context.Context, prNumber int, comment string) error {
-	ctx, _ = r.logger(ctx, "")
-	return r.gh.CreatePullRequestComment(ctx, r.owner, r.repo, prNumber, comment)
-}
-
 // ResourceURLs implements RepositoryWithURLs.
 func (r *githubRepository) ResourceURLs(ctx context.Context, file *FileInfo) (*provisioning.ResourceURLs, error) {
 	cfg := r.config.Spec.GitHub
@@ -777,146 +657,6 @@ func (r *githubRepository) ResourceURLs(ctx context.Context, file *FileInfo) (*p
 	}
 
 	return urls, nil
-}
-
-func (r *githubRepository) createWebhook(ctx context.Context) (pgh.WebhookConfig, error) {
-	secret, err := uuid.NewRandom()
-	if err != nil {
-		return pgh.WebhookConfig{}, fmt.Errorf("could not generate secret: %w", err)
-	}
-
-	cfg := pgh.WebhookConfig{
-		URL:         r.webhookURL,
-		Secret:      secret.String(),
-		ContentType: "json",
-		Events:      subscribedEvents,
-		Active:      true,
-	}
-
-	hook, err := r.gh.CreateWebhook(ctx, r.owner, r.repo, cfg)
-	if err != nil {
-		return pgh.WebhookConfig{}, err
-	}
-
-	// HACK: GitHub does not return the secret, so we need to update it manually
-	hook.Secret = cfg.Secret
-
-	logging.FromContext(ctx).Info("webhook created", "url", cfg.URL, "id", hook.ID)
-	return hook, nil
-}
-
-// updateWebhook checks if the webhook needs to be updated and updates it if necessary.
-// if the webhook does not exist, it will create it.
-func (r *githubRepository) updateWebhook(ctx context.Context) (pgh.WebhookConfig, bool, error) {
-	if r.config.Status.Webhook == nil || r.config.Status.Webhook.ID == 0 {
-		hook, err := r.createWebhook(ctx)
-		if err != nil {
-			return pgh.WebhookConfig{}, false, err
-		}
-		return hook, true, nil
-	}
-
-	hook, err := r.gh.GetWebhook(ctx, r.owner, r.repo, r.config.Status.Webhook.ID)
-	switch {
-	case errors.Is(err, pgh.ErrResourceNotFound):
-		hook, err := r.createWebhook(ctx)
-		if err != nil {
-			return pgh.WebhookConfig{}, false, err
-		}
-		return hook, true, nil
-	case err != nil:
-		return pgh.WebhookConfig{}, false, fmt.Errorf("get webhook: %w", err)
-	}
-
-	hook.Secret = r.config.Status.Webhook.Secret // we always random gen this, so don't use it for mustUpdate below.
-
-	var mustUpdate bool
-
-	if hook.URL != r.webhookURL {
-		mustUpdate = true
-		hook.URL = r.webhookURL
-	}
-
-	if !slices.Equal(hook.Events, subscribedEvents) {
-		mustUpdate = true
-		hook.Events = subscribedEvents
-	}
-
-	if !mustUpdate {
-		return hook, false, nil
-	}
-
-	// Something has changed in the webhook. Let's rotate the secret as well, so as to ensure we end up with a 100% correct webhook.
-	secret, err := uuid.NewRandom()
-	if err != nil {
-		return pgh.WebhookConfig{}, false, fmt.Errorf("could not generate secret: %w", err)
-	}
-	hook.Secret = secret.String()
-
-	if err := r.gh.EditWebhook(ctx, r.owner, r.repo, hook); err != nil {
-		return pgh.WebhookConfig{}, false, fmt.Errorf("edit webhook: %w", err)
-	}
-
-	return hook, true, nil
-}
-
-func (r *githubRepository) deleteWebhook(ctx context.Context) error {
-	if r.config.Status.Webhook == nil {
-		return fmt.Errorf("webhook not found")
-	}
-
-	id := r.config.Status.Webhook.ID
-
-	if err := r.gh.DeleteWebhook(ctx, r.owner, r.repo, id); err != nil {
-		return fmt.Errorf("delete webhook: %w", err)
-	}
-
-	logging.FromContext(ctx).Info("webhook deleted", "url", r.config.Status.Webhook.URL, "id", id)
-	return nil
-}
-
-func (r *githubRepository) OnCreate(ctx context.Context) (*provisioning.WebhookStatus, error) {
-	if len(r.webhookURL) == 0 {
-		return nil, nil
-	}
-
-	ctx, _ = r.logger(ctx, "")
-	hook, err := r.createWebhook(ctx)
-	if err != nil {
-		return nil, err
-	}
-	return &provisioning.WebhookStatus{
-		ID:               hook.ID,
-		URL:              hook.URL,
-		Secret:           hook.Secret,
-		SubscribedEvents: hook.Events,
-	}, nil
-}
-
-func (r *githubRepository) OnUpdate(ctx context.Context) (*provisioning.WebhookStatus, error) {
-	if len(r.webhookURL) == 0 {
-		return nil, nil
-	}
-	ctx, _ = r.logger(ctx, "")
-	hook, _, err := r.updateWebhook(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	return &provisioning.WebhookStatus{
-		ID:               hook.ID,
-		URL:              hook.URL,
-		Secret:           hook.Secret,
-		SubscribedEvents: hook.Events,
-	}, nil
-}
-
-func (r *githubRepository) OnDelete(ctx context.Context) error {
-	if len(r.webhookURL) == 0 {
-		return nil
-	}
-	ctx, _ = r.logger(ctx, "")
-	return r.deleteWebhook(ctx)
 }
 
 func (r *githubRepository) Clone(ctx context.Context, opts CloneOptions) (ClonedRepository, error) {
