+++ pkg/registry/apis/provisioning/webhooks/testdata/webhook-ping-check.json
